refactor: move js/ -> src/
