+++ src/ReactCheatSheet.js
